Refactor [#9] Task 관련 패키지 이동 + Fix #15 conflict resolve
application으로 Service 이동
