--- morib/src/main/java/org/morib/server/api/homeViewApi/service/fetch/task/FetchTaskService.java
@@ -1,6 +1,0 @@
-package org.morib.server.api.homeViewApi.service.fetch.task;
-
-
-public interface FetchTaskService {
-    void fetch();
-}
