--- morib/src/main/java/org/morib/server/api/homeViewApi/service/fetch/task/FetchTaskServiceImpl.java
@@ -1,16 +1,0 @@
-package org.morib.server.api.homeViewApi.service.fetch.task;
-
-import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.task.application.TaskGateway;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class FetchTaskServiceImpl implements FetchTaskService {
-    private final TaskGateway taskGateway;
-
-    @Override
-    public void fetch() {
-
-    }
-}
