+++ morib/src/main/java/org/morib/server/domain/task/application/ToggleTaskStatusService.java
@@ -1,4 +1,4 @@
-package org.morib.server.api.homeViewApi.service.toggle.task;
+package org.morib.server.domain.task.application;
 
 public interface ToggleTaskStatusService {
     void toggle();
