+++ morib/src/main/java/org/morib/server/domain/task/application/ToggleTaskStatusServiceImpl.java
@@ -1,4 +1,4 @@
-package org.morib.server.api.homeViewApi.service.toggle.task;
+package org.morib.server.domain.task.application;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.task.TaskManager;
